feat: implement header according to draft 04
Draft 04 now allows the frame count to be variable in size as well,
as it was possible only with the key_id previously. As a result
the header now can come in 4 flavors, so I took the chance to generalize
the implementation more.

BREAKING CHANGE:
Due to the changes in the draft, the frame count is now serialized differently
if it is < 8. As a result it is no longer compatible with previous drafts
See the (diff)[https://author-tools.ietf.org/iddiff?url1=draft-ietf-sframe-enc-03&url2=draft-ietf-sframe-enc-04&difftype=--html] for details. Also `header::Header` was reimplemented as `header::SframeHeader`.