feat: using meta data as AD
The data skipped for encryption is now used as `meta data`. Together with the serialized header it is used  as AAD used for the authentification (see 4.4.3 in the draft).

BREAKING CHANGE:
As due to this change the AAD used for authenticity protection now differs, encryption/decryption is longer compatible with older library version .