refactor!: introduce a frame counter trait
As of RFC9605 9.1 Applications MUST ensure that each (base_key, KID, CTR) combination is used for at most one SFrame encryption operation.
To make it more obvious for the user this is soft enforced by a trait
instead of passing down the counter directly.

Also renamed FrameCount to Counter to align better with the header
specification